Rectangle Cases Finished
Program can successfully detect whether or not it is a rectangle. Will move onto rectilinear polygons next, the code should be almost identical since you can use the same code for rectangle detection to set it apart from convex or simple polygons. First step should be to modify the getInnerTriangles function to take any number of of points as opposed to the four for rectangles.